refactor(test): split manual backup coverage from save-boundary tests
